Rename all variants of PL/pgSQL in the code and comments for PL/TSQL.
Rename with matching case and conventions.  Supply a description for the
new procedural language where appropriate.  The changeset applies to:

    - Build System Files
    - Source Code Structures
    - Source Code Enumerations
    - Source Code Symbolic Constants
    - Source Code Functions
    - Build System and Source Code Comments
    - Extension Metadata and SQL

Note that this does not include any translation (.po) files.  Many of
the translated messages will be removed or replaced so it does not make
sense to address them presently.