Make the EXTENSION support for pltsql functional.
INSERT into pg_catalog.pg_pltemplate as part of the EXTENSION SQL file
instead of making invasive bootstrap catalog entries in the core.  By
doing so, keep pltsql flexible so it can be installed simply with a
CREATE EXTENSION invocation as a self-contained package.

This does mean, however, that the EXTENSION will require superuser
privilege to install it but the control metadata makes that explicit for
the users at least as far as the corresponding error is concerned.

(To support "CREATE LANGUAGE pltsql" which the EXTENSION uses under the
covers, we need to either manually CREATE the corresponding language
support functions (call handler, validator, ...) and use the
corresponding long CREATE LANGUAGE form OR make an entry into
pg_pltemplate to be able to skip the support-function clauses while
creating the LANGUAGE.

Both require superuser privileges: language handler functions are
UNTRUSTED (C functions) and pg_pltemplate is a system catalog.)